Improve architecture locality
